refactor ContentIndex cache to be wrapped with content-index addon
